課題、ログイン認証追加完了

+++ app/Http/Controllers/TasksController.php
@@ -15,11 +15,17 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
-        
-        return view('tasks.index', [
-            'tasks' => $tasks,
-            ]);
+        $data = [];
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(5);
+            
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        return view('welcome', $data);
     }
 
     /**
@@ -49,11 +55,11 @@ class TasksController extends Controller
             'status' => 'required|max:10',
             ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content, 'status' => $request->status
+        ]);
         
+        // return back();じゃなくてリダイレクトにしてみる
         return redirect('/');
     }
 
@@ -66,10 +72,17 @@ class TasksController extends Controller
     public function show($id)
     {
         $task = Task::find($id);
+        // もしくは、= \App\Task::find($id);
         
-        return view('tasks.show', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.show', [
             'task' => $task,
             ]);
+        }
+        
+        else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -82,9 +95,15 @@ class TasksController extends Controller
     {
         $task = Task::find($id);
         
-        return view('tasks.edit', [
+        if (\Auth::id() === $task->user_id) {
+            return view('tasks.edit', [
             'task' => $task,
             ]);
+        }
+        
+        else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -102,11 +121,16 @@ class TasksController extends Controller
             ]);
         
         $task = Task::find($id);
+        if (\Auth::id() === $task->user_id) {
         $task->content = $request->content;
         $task->status = $request->status;
         $task->save();
         
         return redirect('/');
+        }
+        else {
+            return redirect('/');
+        }
     }
 
     /**
@@ -118,8 +142,13 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::find($id);
-        $task->delete();
-        
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            
+            return redirect('/');
+        }
+        else {
         return redirect('/');
+        }
     }
 }
